fix template files copying

+++ src/common/utils/fs.utils.ts
@@ -3,16 +3,14 @@ import path from 'path';
 
 const copyFolderSync = (from: string, to: string, initialpath: string) => {
     try {
-        if (!from.toLowerCase().includes('node_modules') && !from.toLowerCase().includes('.gitignore') && !from.toLowerCase().includes('dist')) {
-            if (!fs.existsSync(to)) fs.mkdirSync(to); 
-            fs.readdirSync(from).forEach(element => {
-                if (fs.lstatSync(path.join(from, element)).isFile()) {
-                    fs.copyFileSync(path.join(from, element), path.join(to, element));
-                } else {
-                    copyFolderSync(path.join(from, element), path.join(to, element), initialpath);
-                }
-            });
-        }
+        if (!fs.existsSync(to)) fs.mkdirSync(to); 
+        fs.readdirSync(from).forEach(element => {
+            if (fs.lstatSync(path.join(from, element)).isFile() && !from.toLowerCase().includes('.gitignore')) {
+                fs.copyFileSync(path.join(from, element), path.join(to, element));
+            } else {
+                copyFolderSync(path.join(from, element), path.join(to, element), initialpath);
+            }
+        });
     } catch (error) {
         fs.rmdirSync(initialpath);
         process.exit();
